Separate components into small parts
