+++ src/components/product/ProductCard.module.css
@@ -101,47 +101,12 @@
     margin-right: 0.5rem;
 }
 
-.bidder-img {
-    width: 2.75rem;
-    height: 2.75rem;
-    border: 1px solid var(--green);
-    border-radius: 999px;
-    background-color: var(--sub-secondary);
-    grid-row: 1 / span 2;
-    position: relative;
-}
-
-.total-bid {
-    width: 1.5rem;
-    height: 1.5rem;
-    position: absolute;
-    bottom: -0.5rem;
-    right: -0.5rem;
-    border: 1px solid var(--green);
-    border-radius: 999px;
-    background-color: var(--sub-primary);
-    align-self: center;
-    text-align: center;
-    font-size: 0.8125em;
-    line-height: 1.5rem;
-}
-
 .buy {
     display: grid;
     grid-template-columns: 2.5rem 6.75rem;
     border-left: 1px solid var(--sub-secondary);
     padding-left: 0.5rem;
     text-align: right;
-}
-
-.bidder > .headings {
-    font-size: 0.875em;
-    color: var(--green);
-}
-
-.buy > .headings {
-    font-size: 0.875em;
-    color: var(--red);
 }
 
 .price {
@@ -152,39 +117,6 @@
 .icon {
     color: var(--red);
     grid-row: 1 / span 2;
-}
-
-.btn-bid {
-    width: 18.75rem;
-    height: 2.8125rem;
-    border: 2px solid var(--primary);
-    background-color: var(--sub-primary);
-    color: var(--primary);
-    font-weight: 700;
-    font-size: 1.125em;
-    text-transform: uppercase;
-    border-radius: 10px;
-    margin-top: 0.25rem;
-    box-shadow: 0 4px 10px 2px rgba(0, 0, 0, 0.25);
-}
-
-.btn-bid:hover {
-    background-color: var(--primary);
-    color: var(--sub-primary);
-}
-
-.btn-bid span {
-    margin-left: 0.5rem;
-}
-
-.bid-icon {
-    align-self: center;
-    color: var(--primary);
-    margin-left: 0.25;
-}
-
-.btn-bid:hover .bid-icon {
-    color: var(--sub-primary);
 }
 
 .card-bottom {
@@ -233,18 +165,6 @@
         grid-template-columns: 2rem 4.75rem;
     }
 
-    .bidder-img {
-        width: 2rem;
-        height: 2rem;
-    }
-
-    .total-bid {
-        width: 1.25rem;
-        height: 1.25rem;
-        line-height: 1.25rem;
-        font-size: 0.75em;
-    }
-
     .buy {
         grid-template-columns: 2rem 5rem;
     }
@@ -256,16 +176,5 @@
 
     .price {
         font-size: 0.8em;
-    }
-
-    .btn-bid {
-        width: 15rem;
-        height: 2rem;
-        font-size: 0.95em;
-    }
-
-    .option-icon {
-        width: 1.25rem;
-        height: 1.25rem;
     }
 }
