+++ src/components/product/ProductCard.tsx
@@ -1,7 +1,10 @@
 import { Icon } from "@iconify/react";
 import * as React from "react";
 import { Product } from "../../models";
+import Bidder from "./bid-info/Bidder";
+import BidButton from "./button/BidButton";
 import ProductModal from "./modal/product/ProductBidModal";
+import Heading from "./price-heading/Heading";
 import classes from "./ProductCard.module.css";
 import ProductOptions from "./ProductOptions";
 
@@ -28,23 +31,6 @@ const DUMMY_DATA = {
 export default function ProductCard() {
     const [showBidModal, setShowBidModal] = React.useState(false);
     const [isNewProd, setIsNewProd] = React.useState(true);
-
-    const formatPrice = React.useCallback((price: number) => {
-        const lookup = [
-            { value: 1, symbol: "" },
-            { value: 1e3, symbol: "k" },
-            { value: 1e6, symbol: "M" },
-            { value: 1e9, symbol: "B" },
-            { value: 1e12, symbol: "T" },
-        ];
-        const item = lookup
-            .slice()
-            .reverse()
-            .find(function (item) {
-                return price >= item.value;
-            });
-        return item ? (price / item.value).toFixed(1) + item.symbol + "VND" : "N/A";
-    }, []);
 
     const showBidModalHandler = () => {
         setShowBidModal(true);
@@ -77,25 +63,18 @@ export default function ProductCard() {
                     </div>
                     <div className={classes["bid-info"]}>
                         <div className={classes["bidder"]}>
-                            <div className={classes["bidder-img"]}>
-                                <div className={classes["total-bid"]}>
-                                    {DUMMY_DATA.totalBidCount <= 99 ? DUMMY_DATA.totalBidCount : "99+"}
-                                </div>
-                            </div>
-                            <div className={classes["headings"]}>Giá Hiện Tại</div>
+                            <Bidder totalBidCount={DUMMY_DATA.totalBidCount} />
+                            <Heading content="Giá hiện tại" color="#6fc47f" />
                             <div className={classes.price}>{1}</div>
                         </div>
                         <div className={classes["buy"]}>
                             <Icon icon="emojione-monotone:money-bag" className={classes.icon} width={42} height={45} />
-                            <div className={classes["headings"]}>Mua Ngay</div>
+                            <Heading content="Mua ngay" color="#ee4730" />
                             <div className={classes.price}>{1}</div>
                         </div>
                     </div>
                     <div className={classes["card-bottom"]}>
-                        <button className={classes["btn-bid"]} onClick={showBidModalHandler}>
-                            <Icon icon="ri:auction-fill" width={24} height={24} className={classes["bid-icon"]} />
-                            <span>Ra giá</span>
-                        </button>
+                        <BidButton openModal={showBidModalHandler} />
                         <ProductOptions />
                     </div>
                 </div>
