+++ src/components/product/bid-info/Bidder.module.css
@@ -1,0 +1,38 @@
+.bidder-img {
+    width: 2.75rem;
+    height: 2.75rem;
+    border: 1px solid var(--green);
+    border-radius: 999px;
+    background-color: var(--sub-secondary);
+    grid-row: 1 / span 2;
+    position: relative;
+}
+
+.total-bid {
+    width: 1.5rem;
+    height: 1.5rem;
+    position: absolute;
+    bottom: -0.5rem;
+    right: -0.5rem;
+    border: 1px solid var(--green);
+    border-radius: 999px;
+    background-color: var(--sub-primary);
+    align-self: center;
+    text-align: center;
+    font-size: 0.8125em;
+    line-height: 1.5rem;
+}
+
+@media screen and (max-width: 1400px) {
+    .bidder-img {
+        width: 2rem;
+        height: 2rem;
+    }
+
+    .total-bid {
+        width: 1.25rem;
+        height: 1.25rem;
+        line-height: 1.25rem;
+        font-size: 0.75em;
+    }
+}
