+++ src/components/product/button/BidButton.module.css
@@ -1,0 +1,45 @@
+.btn-bid {
+    width: 18.75rem;
+    height: 2.8125rem;
+    border: 2px solid var(--primary);
+    background-color: var(--sub-primary);
+    color: var(--primary);
+    font-weight: 700;
+    font-size: 1.125em;
+    text-transform: uppercase;
+    border-radius: 10px;
+    margin-top: 0.25rem;
+    box-shadow: 0 4px 10px 2px rgba(0, 0, 0, 0.25);
+}
+
+.btn-bid:hover {
+    background-color: var(--primary);
+    color: var(--sub-primary);
+}
+
+.btn-bid span {
+    margin-left: 0.5rem;
+}
+
+.bid-icon {
+    align-self: center;
+    color: var(--primary);
+    margin-left: 0.25;
+}
+
+.btn-bid:hover .bid-icon {
+    color: var(--sub-primary);
+}
+
+@media screen and (max-width: 1400px) {
+    .btn-bid {
+        width: 15rem;
+        height: 2rem;
+        font-size: 0.95em;
+    }
+
+    .bid-icon {
+        width: 1.25rem;
+        height: 1.25rem;
+    }
+}
