+++ src/components/product/price-heading/Heading.module.css
@@ -1,0 +1,10 @@
+.heading {
+    font-size: 0.875em;
+    text-transform: capitalize;
+}
+
+@media screen and (max-width: 1400px) {
+    .heading {
+        font-size: 0.75em;
+    }
+}
